Added help text to all bash scripts
Committer: Ntipac <ntch98@gmail.com>
On branch help_text
Changes to be committed:
	modified:   assign_learners.sh
	modified:   channel_setup/create_channel_module.sh
	modified:   channel_setup/delete_channels.sh
	modified:   channel_setup/delete_prereqs.sh
	modified:   channel_setup/export_channels.sh
	modified:   channel_setup/import_channels.sh
	modified:   channel_setup/import_channels_network.sh
	modified:   channel_setup/insert_channel_subscriptions.sh
	modified:   channel_setup/reorder_channels.sh
	modified:   restore_backups.sh

+++ channel_setup/delete_prereqs.sh
@@ -1,3 +1,17 @@
 #!/bin/bash
+
+
+if [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
+  echo "NAME"
+  echo "  delete_prereqs - Delete channel prerequisites"
+  echo
+  echo "DESCRIPTION"
+  echo "  Deletes all numeracy channel prerequisites"
+  echo 
+  echo "Example"
+  echo " ./delete_prereqs.sh"
+  exit 1
+fi
+
 # Delete all content prerequisites
 PGPASSWORD=$KOLIBRI_DATABASE_PASSWORD psql -h "$KOLIBRI_DATABASE_HOST" -U "$KOLIBRI_DATABASE_USER" -d "$KOLIBRI_DATABASE_NAME" -p "$KOLIBRI_DATABASE_PORT" -c "DELETE FROM content_contentnode_has_prerequisite;"
